commits

+++ app/Http/Controllers/ListingController.php
@@ -3,12 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Models\Listing;
-use Illuminate\Validation\Rule;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 
 class ListingController extends Controller
 {
-
     // Show all listings
     public function index()
     {
@@ -17,7 +16,7 @@ class ListingController extends Controller
         ]);
     }
 
-    // Show single listing
+    //Show single listing
     public function show(Listing $listing)
     {
         return view('listings.show', [
@@ -25,11 +24,12 @@ class ListingController extends Controller
         ]);
     }
 
-    // Show create form
+    // Show Create Form
     public function create()
     {
         return view('listings.create');
     }
+
     // Store Listing Data
     public function store(Request $request)
     {
@@ -37,42 +37,45 @@ class ListingController extends Controller
             'title' => 'required',
             'company' => ['required', Rule::unique('listings', 'company')],
             'location' => 'required',
-            'location' => 'required',
             'website' => 'required',
             'email' => ['required', 'email'],
             'tags' => 'required',
-            'description' => 'required',
+            'description' => 'required'
         ]);
 
         if ($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
-
         $formFields['user_id'] = auth()->id();
 
         Listing::create($formFields);
-        return redirect('/')->with('message', 'Listing created successfuly');
+
+        return redirect('/')->with('message', 'Listing created successfully!');
     }
 
-    // show edit form
+    // Show Edit Form
     public function edit(Listing $listing)
     {
         return view('listings.edit', ['listing' => $listing]);
     }
 
-    // update listing
+    // Update Listing Data
     public function update(Request $request, Listing $listing)
     {
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
             'location' => 'required',
-            'location' => 'required',
             'website' => 'required',
             'email' => ['required', 'email'],
             'tags' => 'required',
-            'description' => 'required',
+            'description' => 'required'
         ]);
 
         if ($request->hasFile('logo')) {
@@ -80,13 +83,25 @@ class ListingController extends Controller
         }
 
         $listing->update($formFields);
-        return back()->with('message', 'Listing updated successfuly');
+
+        return back()->with('message', 'Listing updated successfully!');
     }
 
-    // delete listing
+    // Delete Listing
     public function destroy(Listing $listing)
     {
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
-        return redirect('/')->with('message', 'Lisiting Deleted Successfully');
+        return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    // Manage Listings
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings]);
     }
 }
